Take absolute value of net metered channel
Previous versions of this integration were taking the polarized value
when net metering was on. Switching to taking the net value resulted in
some installations starting to report negative values. While that may
have been technically correct, Home Assistant expects positive values to
be flowing into the energy dashboard, so we'll take the absolute value.